add tv80 core
this is a fork of the TV80 project created by Guy Hutchison and Howard M. Harte. we made a number of modifications and bug fixes to make this work with the GameBoy's instruction set, but those changes should probably be revisted because it's been a few years. eventually this should be moved to its own repository.